# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dastyare Social — CS",
    "version": "1.0.0",
    "description": "REST API for Dastyare Social — Creator Studio. Manage posts (threads, shorts, media) and stories. Supports JSON and multipart uploads, reactions, views, and likes. Authentication via Your API key. See /llms.txt for LLM-oriented documentation."
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreatePostBody": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "nullable": true
          },
          "media": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "text",
                    "image",
                    "video",
                    "voice",
                    "file"
                  ],
                  "nullable": true
                },
                "width": {
                  "type": "number"
                },
                "height": {
                  "type": "number"
                },
                "duration": {
                  "type": "number"
                }
              },
              "required": [
                "url"
              ]
            }
          }
        }
      },
...
      "MediaUploadResponse": {},
      "patchPostsSchema": {
        "type": "object"
      },
...
      "PostsCountResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          }
        },
        "required": [
          "total"
        ]
      },
      "PostsListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {}
          },
          "total": {
            "type": "number"
          },
          "hasMore": {
            "type": "boolean"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "hasMore",
          "page",
          "limit"
        ]
      },
...
      "PostsResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/PostsListResponse"
          },
          {
            "$ref": "#/components/schemas/PostsCountResponse"
          }
        ]
      },
      "PostSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "PostWithReactionsSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "video",
              "voice",
              "file"
            ]
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "views": {
            "type": "string"
          },
          "pinnedAt": {
            "type": "string",
            "nullable": true
          },
          "media": {},
          "createdAt": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "reactions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "emoji": {
                  "type": "string"
                },
                "count": {
                  "type": "number"
                }
              },
              "required": [
                "emoji",
                "count"
              ]
            }
          }
        },
        "required": [
          "id",
          "type",
          "content",
          "views",
          "pinnedAt",
          "media",
          "createdAt",
          "updatedAt",
          "reactions"
        ]
      },
...
    },
...
  },
  "paths": {
    "/upload": {
      "post": {
        "operationId": "post-upload",
        "summary": "Upload Media — Single File",
        "description": "Single-file upload endpoint (multipart/form-data with a `file` field). Server-side upload — buffers the file then uploads to S3. Returns media metadata. For real upload progress, use `/api/upload/presign` + `/api/upload/confirm` (direct browser-to-S3) instead.",
        "tags": [
          "Media"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "x-jsonschema-contentMediaType": "application/octet-stream"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "description": "Multipart form data containing a file upload.",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponse"
                },
                "examples": {
                  "example": {
                    "value": {
                      "url": "https://cdn.example.com/media/image/abc.jpg",
                      "key": "media/image/abc.jpg",
                      "kind": "image",
                      "mimeType": "image/jpeg",
                      "size": 12345,
                      "width": 1080,
                      "height": 1920,
                      "duration": 0,
                      "filename": "photo.jpg"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/upload-stream": {
      "post": {
        "operationId": "post-upload-stream",
        "summary": "Upload Media — Streaming with Progress",
        "description": "Upload a single media file with server-side progress tracking.",
        "tags": [
          "Media"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "x-jsonschema-contentMediaType": "application/octet-stream"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "description": "Multipart form data containing a file upload.",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/200 - SSE stream with progress/done/error events"
                }
              }
            }
          },
          "400 - Invalid request": {
            "$ref": "#/components/responses/400 - Invalid request"
          },
          "403 - Demo mode active": {
            "$ref": "#/components/responses/403 - Demo mode active"
          },
          "500 - Server error": {
            "$ref": "#/components/responses/500 - Server error"
          }
        }
      }
    },
...
    "/posts": {
      "get": {
        "operationId": "get-posts",
        "summary": "List Posts — Paginated Feed",
        "description": "Returns paginated posts with reactions. Use query type=count for total count, type=shorts for vertical videos (1080x1920). Default type=list.",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "example"
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "example"
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "example"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-posts",
        "summary": "Create Post — Upload Media Via URL",
        "description": "Create a post from a JSON body. Media must be referenced by URL — upload files to `/api/upload` or `/api/upload-stream` first, then map the returned `url`, `kind` (as `type`), `width`, `height`, and `duration` into the `media` array below. JSON body with action=batch-view and ids array increments views for multiple posts.",
        "tags": [
          "Posts"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostBody"
              },
              "examples": {
                "example": {
                  "value": {
                    "content": "Hello",
                    "media": [
                      {
                        "url": "https://cdn.example.com/media/image/abc.jpg",
                        "type": "image",
                        "width": 1080,
                        "height": 1920
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostWithReactionsSchema"
                }
              }
            }
          },
          "PostSuccessResponse": {
            "$ref": "#/components/responses/PostSuccessResponse"
          }
        }
      }
    },
    "/posts/{post_id}": {
      "get": {
        "operationId": "get-posts-{post_id}",
        "summary": "Read Post — Get by ID",
        "description": "Returns a single post with reactions.",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostWithReactionsSchema"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-posts-{post_id}",
        "summary": "Post Actions — React or View",
        "description": "Perform actions on a post. action=reaction requires emoji string. action=view increments the view count.",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostWithReactionsSchema"
                }
              }
            }
          },
          "PostSuccessResponse": {
            "$ref": "#/components/responses/PostSuccessResponse"
          }
        }
      },
      "patch": {
        "operationId": "patch-posts-{post_id}",
        "summary": "Update Post — Edit Fields",
        "description": "Partial update of post fields (content, views, pinnedAt, media, type).",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/patchPostsSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostWithReactionsSchema"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-posts-{post_id}",
        "summary": "Delete Post — Remove by ID",
        "description": "Permanently deletes a post by ID.",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostSuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "API server"
    }
  ]
}